Moved MixedFeed.tsx and FilteredFeed.tsx to Explore folder
